requirements.txt
numpy
scipy
pyzmq
gnuradio
scikit-learn
